🐳 chore: 新增类型导出

+++ build/get-dts-plugin.mjs
@@ -2,18 +2,20 @@ import fs from 'node:fs'
 import { spawn } from 'node:child_process'
 import dtsRolldown from 'unplugin-dts/rolldown'
 
-export function dts(scopePkgName) {
+export function dts(scopePkgName, bundleTypes = true) {
   return dtsRolldown({
     afterBuild() {
-      const p = spawn('npx', ['api-extractor', 'run', '--local'], { stdio: 'inherit' })
-      p.on('close', code => {
-        if (code !== 0) {
-          console.error(`API Extractor process exited with code ${code}`)
-        } else {
-          console.log('API Extractor completed successfully')
-          fs.rmdirSync(`./dist/${scopePkgName}`, { recursive: true, force: true })
-        }
-      })
+      if (bundleTypes) {
+        const p = spawn('npx', ['api-extractor', 'run', '--local'], { stdio: 'inherit' })
+        p.on('close', code => {
+          if (code !== 0) {
+            console.error(`API Extractor process exited with code ${code}`)
+          } else {
+            console.log('API Extractor completed successfully')
+            fs.rmdirSync(`./dist/${scopePkgName}`, { recursive: true, force: true })
+          }
+        })
+      }
     },
   })
 }
